pyproject.toml
[project]
name = "lg-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "langchain>=1.2.6",
    "langchain-google-genai>=4.2.0",
    "langgraph>=1.0.6",
    "langsmith>=0.6.4",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
